fix(console): preserve desktop protocol exports
