+++ runtime/fleet-console/desktop-protocol/index.ts
@@ -33,20 +33,6 @@ export interface ResolveCanonicalLocalConsolePathsInput {
   readonly packageRoot: string;
 }
 
-export type ConsoleThemeId = "instrument" | "maritime" | "carbon";
-export type DesktopThemeId = string;
-
-export interface DesktopTitleBarOverlay {
-  readonly color: string;
-  readonly symbolColor: string;
-  readonly height: number;
-}
-
-export interface DesktopThemeSnapshot {
-  readonly theme: DesktopThemeId;
-  readonly titleBarOverlay: DesktopTitleBarOverlay;
-}
-
 export const DESKTOP_PROTOCOL_VERSION = 1;
 export const DESKTOP_RESOURCE_ROOT_ENV = "FLEET_CONSOLE_RESOURCE_ROOT";
 export const DESKTOP_OWNER_ID_ENV = "FLEET_CONSOLE_OWNER_ID";
@@ -54,9 +40,6 @@ export const DESKTOP_OWNER_KIND_ENV = "FLEET_CONSOLE_OWNER_KIND";
 export const DESKTOP_PROTOCOL_VERSION_ENV = "FLEET_CONSOLE_PROTOCOL_VERSION";
 export const DESKTOP_RESOURCE_ROOT_MARKER = ".fleet-console-resource-root";
 export const DESKTOP_DEVELOPMENT_ENV = "FLEET_CONSOLE_DESKTOP_DEVELOPMENT";
-export const DESKTOP_THEME_PATH = "/api/v1/desktop/theme";
-export const DESKTOP_THEME_EVENTS_PATH = "/api/v1/desktop/theme/events";
-export const DESKTOP_THEME_EVENT = "desktop:theme";
 
 const LOCK_DIR_NAME = "fleet-console";
 const LOCK_FILE_NAME = "console.lock";
@@ -64,10 +47,6 @@ const CONSOLE_DATA_DIR_NAME = "console";
 const CONSOLE_STATE_FILE_NAME = "state.json";
 const CONSOLE_SETTINGS_FILE_NAME = "settings.json";
 const CONSOLE_CAPTURES_DIR_NAME = "captures";
-const ELECTRON_COLOR_PATTERN = /^#(?:[\da-f]{3,4}|[\da-f]{6}|[\da-f]{8})$/i;
-const MIN_TITLE_BAR_OVERLAY_HEIGHT = 24;
-const MAX_TITLE_BAR_OVERLAY_HEIGHT = 128;
-const DESKTOP_THEME_ID_PATTERN = /^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$/;
 
 export function isDesktopDevelopmentEnvironment(env: Readonly<Record<string, string | undefined>>): boolean {
   return env[DESKTOP_DEVELOPMENT_ENV] === "1";
@@ -111,30 +90,4 @@ export function formatDesktopResourceRootMarker(): string {
 
 export function isDesktopResourceRootMarkerValid(content: string): boolean {
   return content.trim() === String(DESKTOP_PROTOCOL_VERSION);
-}
-
-export function isDesktopThemeSnapshot(value: unknown): value is DesktopThemeSnapshot {
-  if (!isRecord(value) || !isDesktopThemeId(value.theme) || !isRecord(value.titleBarOverlay)) return false;
-  return isElectronColor(value.titleBarOverlay.color)
-    && isElectronColor(value.titleBarOverlay.symbolColor)
-    && isTitleBarOverlayHeight(value.titleBarOverlay.height);
-}
-
-function isDesktopThemeId(value: unknown): value is DesktopThemeId {
-  return typeof value === "string" && DESKTOP_THEME_ID_PATTERN.test(value);
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-function isElectronColor(value: unknown): value is string {
-  return typeof value === "string" && ELECTRON_COLOR_PATTERN.test(value);
-}
-
-function isTitleBarOverlayHeight(value: unknown): value is number {
-  return typeof value === "number"
-    && Number.isInteger(value)
-    && value >= MIN_TITLE_BAR_OVERLAY_HEIGHT
-    && value <= MAX_TITLE_BAR_OVERLAY_HEIGHT;
 }
