+++ runtime/fleet-console/tests/desktop-protocol.test.ts
@@ -5,7 +5,7 @@ import { fileURLToPath } from "node:url";
 
 import { afterEach, describe, expect, it } from "vitest";
 
-import { formatDesktopResourceRootMarker, isDesktopResourceRootMarkerValid, isDesktopThemeSnapshot } from "@fleet-console/desktop-protocol";
+import { formatDesktopResourceRootMarker, isDesktopResourceRootMarkerValid } from "@fleet-console/desktop-protocol";
 
 import {
   DESKTOP_DEVELOPMENT_ENV,
@@ -46,14 +46,10 @@ describe("desktop protocol", () => {
     expect(source).not.toContain("desktopThemeSnapshot");
   });
 
-  it("accepts safely shaped future Console overlay colors without bundling a palette", () => {
-    expect(isDesktopThemeSnapshot({ theme: "maritime", titleBarOverlay: { color: "#123456", symbolColor: "#abcdef", height: 48 } })).toBe(true);
-    expect(isDesktopThemeSnapshot({ theme: "future-aurora", titleBarOverlay: { color: "#123456", symbolColor: "#abcdef", height: 48 } })).toBe(true);
-    expect(isDesktopThemeSnapshot({ theme: "carbon", titleBarOverlay: { color: "#1234", symbolColor: "#abc", height: 52 } })).toBe(true);
-    expect(isDesktopThemeSnapshot({ theme: "future aurora", titleBarOverlay: { color: "#123456", symbolColor: "#abcdef", height: 48 } })).toBe(false);
-    expect(isDesktopThemeSnapshot({ theme: "carbon", titleBarOverlay: { color: "white", symbolColor: "#abcdef", height: 44 } })).toBe(false);
-    expect(isDesktopThemeSnapshot({ theme: "carbon", titleBarOverlay: { color: "#abcdef", symbolColor: "#abcdef", height: 12.5 } })).toBe(false);
-    expect(isDesktopThemeSnapshot({ theme: "carbon", titleBarOverlay: { color: "#abcdef", symbolColor: "#abcdef", height: 129 } })).toBe(false);
+  it("publishes no theme transport exports", () => {
+    const source = fs.readFileSync(path.join(CONSOLE_PACKAGE_ROOT, "desktop-protocol", "index.ts"), "utf8");
+
+    expect(source).not.toMatch(/^export (?:type |interface |const |function )(?:ConsoleThemeId|DesktopThemeId|DesktopTitleBarOverlay|DesktopThemeSnapshot|DESKTOP_THEME_PATH|DESKTOP_THEME_EVENTS_PATH|DESKTOP_THEME_EVENT|isDesktopThemeSnapshot)\b/gmu);
   });
 
   it("validates the marked resource root and exact owner protocol", () => {
