+++ runtime/fleet-console/tests/desktop-theme-routes.test.ts
@@ -2,9 +2,7 @@ import type http from "node:http";
 
 import { describe, expect, it } from "vitest";
 
-import { DESKTOP_THEME_EVENTS_PATH, DESKTOP_THEME_PATH } from "@fleet-console/desktop-protocol";
-
-import { desktopThemeSnapshot } from "../core/host/desktop-theme.js";
+import { DESKTOP_THEME_EVENTS_PATH, DESKTOP_THEME_PATH, desktopThemeSnapshot } from "../core/host/desktop-theme.js";
 import { createDesktopThemeRouter, DESKTOP_THEME_API_CATALOG } from "../core/host/desktop-theme-routes.js";
 
 describe("desktop theme routes", () => {
