+++ runtime/fleet-desktop/src/desktop-theme-sync.ts
@@ -1,10 +1,21 @@
-import {
-  DESKTOP_THEME_EVENT,
-  DESKTOP_THEME_EVENTS_PATH,
-  DESKTOP_THEME_PATH,
-  isDesktopThemeSnapshot,
-  type DesktopThemeSnapshot,
-} from "@fleet-console/desktop-protocol";
+interface DesktopTitleBarOverlay {
+  readonly color: string;
+  readonly symbolColor: string;
+  readonly height: number;
+}
+
+interface DesktopThemeSnapshot {
+  readonly theme: string;
+  readonly titleBarOverlay: DesktopTitleBarOverlay;
+}
+
+const DESKTOP_THEME_PATH = "/api/v1/desktop/theme";
+const DESKTOP_THEME_EVENTS_PATH = "/api/v1/desktop/theme/events";
+const DESKTOP_THEME_EVENT = "desktop:theme";
+const ELECTRON_COLOR_PATTERN = /^#(?:[\da-f]{3,4}|[\da-f]{6}|[\da-f]{8})$/i;
+const MIN_TITLE_BAR_OVERLAY_HEIGHT = 24;
+const MAX_TITLE_BAR_OVERLAY_HEIGHT = 128;
+const DESKTOP_THEME_ID_PATTERN = /^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$/;
 
 export interface DesktopThemeSynchronizer {
   start(origin: string): Promise<void>;
@@ -148,6 +159,32 @@ export function parseDesktopThemeEvent(frame: string): DesktopThemeSnapshot | nu
   } catch {
     return null;
   }
+}
+
+function isDesktopThemeSnapshot(value: unknown): value is DesktopThemeSnapshot {
+  if (!isRecord(value) || !isDesktopThemeId(value.theme) || !isRecord(value.titleBarOverlay)) return false;
+  return isElectronColor(value.titleBarOverlay.color)
+    && isElectronColor(value.titleBarOverlay.symbolColor)
+    && isTitleBarOverlayHeight(value.titleBarOverlay.height);
+}
+
+function isDesktopThemeId(value: unknown): value is string {
+  return typeof value === "string" && DESKTOP_THEME_ID_PATTERN.test(value);
+}
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+
+function isElectronColor(value: unknown): value is string {
+  return typeof value === "string" && ELECTRON_COLOR_PATTERN.test(value);
+}
+
+function isTitleBarOverlayHeight(value: unknown): value is number {
+  return typeof value === "number"
+    && Number.isInteger(value)
+    && value >= MIN_TITLE_BAR_OVERLAY_HEIGHT
+    && value <= MAX_TITLE_BAR_OVERLAY_HEIGHT;
 }
 
 function desktopThemeUrl(origin: string, pathname: string): string {
